+++ CHANGELOG.rst
@@ -1,15 +1,14 @@
 2.0 (unreleased)
 ----------------
 
-- Try and fix DB connection errors when the DB server unexpectedly closes the connection
-- Refactor: use dataclasses
+- Refactor: use dataclasses, all discord infos in extra (including players discord ID)
+- Performance: use orjson, async DB commands and connection pool
 
 1.2 (2023-03-12)
 ----------------
 
 - Fix major player numbering issue (ended up mixing scores and seating in LBL)
 - Improved logging
-
 
 1.1 (2023-02-26)
 ----------------